Add back button for home fragment and post details fragment. Fix bugs

+++ app/src/main/java/com/example/connectue/fragmets/ProfileFragment.java
@@ -503,7 +503,7 @@ public class ProfileFragment extends Fragment {
         StorageReference filePath = FirebaseStorage.getInstance().getReference("posts")
                 .child(emailStr + "." + General.getFileExtension(getContext(), imageUri));
 
-        if (imageURL != null || !imageURL.equals("")) {
+        if (imageURL != null && !imageURL.equals("")) {
             StorageReference currentPicture = FirebaseStorage.getInstance().
                     getReferenceFromUrl(imageURL);
             currentPicture.delete().addOnSuccessListener(new OnSuccessListener<Void>() {
